requirements.txt
asgiref==3.4.1
Django==4.0
django-nested-admin==3.3.3
django-phonenumber-field==6.0.0
django-utils2==2.12.1
mysqlclient==2.1.0
phonenumberslite==8.12.37
Pillow==8.4.0
python-monkey-business==1.0.0
python-utils==2.5.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
